+++ unfurl/tosca_plugins/functions.py
@@ -1,5 +1,5 @@
 """
-This module contains utility functions that can be executed in `"spec" mode <global_state_mode>` (e.g. as part of a class definition or in ``_class_init_``)
+This module contains utility functions that can be executed in `"parse" mode <global_state_mode>` (e.g. as part of a class definition or in ``_class_init_``)
 and in the safe mode Python sandbox.
 Each of these are also available as Eval `Expression Functions`.
 """